refactor(api/v2): extract analytics domain into its own package
Move the analytics domain (the last and largest Phase 6 domain) out of the
internal/api/v2 facade (package api) into a new internal/api/v2/analytics
package, so it builds and runs as its own isolated -race test binary and no
longer gates the per-package -race timeout.

Sub-areas moved (handlers retyped *Controller -> *Handler, bodies verbatim):

- analytics: /analytics/species/*, /analytics/time/* (including the activity
  heatmap), /analytics/confidence/distribution, /analytics/sun,
  /analytics/sources (RegisterAnalyticsRoutes).
- heatmap: the geographic GET /range/heatmap endpoint, the heatmapLRU cache,
  the heatmapSem concurrency semaphore, the singleflight dedup, and the
  classifier.OnRangeFilterReload hook (RegisterHeatmapRoutes).
- insights + dashboard KPIs: /insights/* and /dashboard/kpis, plus the lazily
  built insights repository (RegisterInsightsRoutes).
- database overview: the auth-protected GET /system/database/overview,
  registered from the facade initSystemRoutes at its existing position
  (RegisterDatabaseOverviewRoutes).

Handler embeds *apicore.Core by pointer and is injected the same facade-owned
dependencies as the detections domain: the auth check (isClientAuthenticated)
and the cached name-map read accessors (loadCommonNameMap /
loadCommonToScientificMap). It reuses the already-promoted apicore helpers
(PercentageMultiplier, ParseConfidence/Hour/DateRange filters,
ParsePaginationLimit, StatusClientClosedRequest, NormalizeForLookup,
ResolveSpeciesToScientific) instead of duplicating them.

Name-map plumbing decision: the cached scientific<->common name maps and the
authoritative resolver are shared infrastructure (analytics, the detections
search resolver, the species image handler, settings exclude-list
canonicalization, and the external internal/analysis callers). They stay
facade-owned, relocated from insights.go into a new name_maps.go in package api
(nameMaps/buildNameMaps/loadNameMaps/loadCommonNameMap/loadCommonToScientificMap/
canonicalizeExcludeList/UpdateCommonNameMap/SetNameResolver/loadNameResolver).
This keeps the UpdateCommonNameMap/SetNameResolver external surface on
*Controller unchanged (no signature change for internal/analysis), keeps
settings.go's exclude-list canonicalization working, and avoids any
domain->facade or domain->domain dependency. Analytics gets its own thin
resolveSpeciesToScientific over apicore using the injected accessor; the
facade initInsightsRoutes seeds the maps (gated on V2Manager) before calling
RegisterInsightsRoutes, preserving the original startup behavior.

The analytics-only validation/pagination/hourly-distribution helpers moved out
of the facade utils.go into the analytics package (analytics is their sole
remaining consumer). Tests were rebuilt against a minimal apicore.Core
mirroring the previous setupAnalyticsTestEnvironment, with name-resolution
tests seeding the injected maps directly. The name-map builder tests
(buildNameMaps) stay in package api with the plumbing.

No public-behavior change: apiv2.Controller/New and the external method set are
unchanged, the route-enumeration gate is byte-identical, and hot-reload of the
name maps is observed by analytics via the live atomic snapshot.
