+++ internal/api/v2/analytics/analytics_sources_test.go
@@ -1,4 +1,4 @@
-package api
+package analytics
 
 import (
 	"context"
