+++ internal/api/v2/analytics/analytics_succession_test.go
@@ -1,4 +1,4 @@
-package api
+package analytics
 
 import (
 	"context"
